fix error migrate heroku

+++ database/migrations/2025_11_21_000002_create_product_variant_options_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('product_variant_options')) {
+            return;
+        }
+
         Schema::create('product_variant_options', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_variant_id')->constrained()->onDelete('cascade');
